+++ CHANGES.md
@@ -1,5 +1,9 @@
 # CHANGES
 
+Next Release
+
+- Add `--one-shot` mode to allow copying data as it is from stdin.
+
 v0.2.1
 
 - The first public release.